svn path=/trunk/TheoryMathsExtensions/; revision=15749

+++ org.eventb.theory.ui/src/org/eventb/theory/language/ui/explorer/TheoryPathRootActionProvider.java
@@ -121,7 +121,9 @@ public class TheoryPathRootActionProvider extends CommonActionProvider {
 											theoryPathFile.delete(true, new NullProgressMonitor());
 										}
 									}
-									rodinFile.delete(true, new NullProgressMonitor());
+									if (rodinFile.exists()) {
+										rodinFile.delete(true, new NullProgressMonitor());
+									}
 
 								} catch (PartInitException e) {
 									MessageDialog.openError(null, "Error", "Could not delete file");
